+++ spec/alces_utils_spec.rb
@@ -119,10 +119,6 @@ RSpec.describe AlcesUtils do
         expect(alces.groups.send(group2).name).to eq(group2)
       end
 
-      it 'sets the config to blank' do
-        expect(alces.groups.send(group).config.to_h).to be_empty
-      end
-
       # The mocking would otherwise alter the actual file
       it 'errors if FakeFS is off' do
         FakeFS.deactivate!
@@ -159,11 +155,6 @@ RSpec.describe AlcesUtils do
         expect(alces.node.genders).to eq([AlcesUtils.default_group])
       end
 
-      it 'creates the node with a blank config and answer' do
-        expect(alces.node.config.to_h).to be_empty
-        expect(alces.node.answer.to_h).to be_empty
-      end
-
       it 'errors if the node already exists' do
         expect do
           AlcesUtils.mock(self) { mock_node(name) }
@@ -180,6 +171,34 @@ RSpec.describe AlcesUtils do
           node = alces.nodes.find_by_name(new_node)
           expect(node.genders).to eq(genders)
         end
+      end
+    end
+
+    describe '#reset_alces' do
+      before :each do
+        @old_alces = alces
+        AlcesUtils.mock(self) do
+          config(alces.domain, key: 'I should be deleted in the reset')
+        end
+        reset_alces
+      end
+
+      it 'resets alces to a new instance' do
+        expect(alces).not_to eq(@old_alces)
+      end
+
+      it 'removes the old config mocking' do
+        expect(alces.domain.config.keys).to be_nil
+      end
+
+      it 'sets the Alces.new method to return the new alces object' do
+        new_alces = Metalware::Namespaces::Alces.new
+        expect(new_alces).to eq(alces)
+      end
+
+      it 'returns the new version of alces' do
+        return_from_reset_alces = reset_alces
+        expect(return_from_reset_alces).to eq(alces)
       end
     end
   end