+++ src/question_tree.rb
@@ -2,9 +2,12 @@
 # frozen_string_literal: true
 
 require 'rubytree'
+require 'ostruct'
 
 module Metalware
   class QuestionTree < Tree::TreeNode
+    attr_accessor :answer
+
     BASE_TRAVERSALS = [
       :each,
       :breadth_each,
@@ -19,10 +22,20 @@ module Metalware
       end
     end
 
+    def ask_questions
+      filtered_each.with_index do |question, index|
+        next unless ask_conditional_question?(question)
+        progress = "(#{index + 1}/#{questions_length})"
+        yield create_question(question, progress)
+      end
+    end
+
     def questions_length
-      num = 0
-      filtered_each { |_q| num += 1 }
-      num
+      @questions_length ||= begin
+        num = 0
+        filtered_each { |_q| num += 1 }
+        num
+      end
     end
 
     def question?
@@ -33,8 +46,66 @@ module Metalware
       filtered_each.map(&:identifier)
     end
 
+    # START REMAPPING 'content' to variable names
+    delegate :choices, :optional, :type, to: :os_content
+
     def identifier
-      content[:identifier]
+      os_content.identifier&.to_sym
+    end
+
+    def text
+      os_content.question
+    end
+
+    def yaml_default
+      os_content.default
+    end
+    # END REMAPPING CONTENT
+
+    def section_tree(section)
+      root.children.find { |c| c.name == section }
+    end
+
+    def root_defaults
+      @root_defaults ||= begin
+        section_tree(:domain).filtered_each.reduce({}) do |memo, question|
+          memo.merge(question.identifier => question.yaml_default)
+        end
+      end
+    end
+
+    def flatten
+      filtered_each.reduce({}) do |memo, node|
+        memo.merge(node.identifier => node)
+      end
+    end
+
+    private
+
+    # TODO: Stop wrapping the content in the validator, that should really
+    # be done within the QuestionTree object. It doesn't hurt, as you can't
+    # double wrap and OpenStruct, it just isn't required
+    def os_content
+      OpenStruct.new(content)
+    end
+
+    # NOTE: The following methods are used by the iterator and thus do not
+    # reference the self object
+    def ask_conditional_question?(question)
+      # Ask the question if the parent has a truthy answer
+      if question.parent.answer
+        true
+      # Ask the question if the parent isn't a question (e.g. a section)
+      elsif !question.parent.question?
+        true
+      # Otherwise don't ask the question
+      else
+        false
+      end
+    end
+
+    def create_question(question, progress_indicator)
+      Configurator::Question.new(question, progress_indicator)
     end
   end
 end